Security Changes

+++ src/main/java/com/tool/RecruitXpert/Repository/RecruiterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecruiterRepository extends JpaRepository<Recruiter,Long> {
+public interface RecruiterRepository extends JpaRepository<Recruiter,Integer> {
 }
